simplified user loggin eventing
