+++ frontend/components/TagList.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+const TagList: React.FC = () => {
+    return ( <div> tagList</div>)
+}
+
+export default TagList;
