+++ frontend/services/auth.service.tsx
@@ -1,7 +1,6 @@
 import axios from "axios";
 import { alertService } from "./alert.service";
 import { credentials } from "../app/account/login/page";
-import useStorage from '@components/useStorage'
 import React from "react";
 export interface User {
   username: string;
